+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapViewController.swift
@@ -841,7 +841,6 @@ extension PXOneTapViewController: MLCardFormLifeCycleDelegate {
 extension PXOneTapViewController: UINavigationControllerDelegate {
     func navigationController(_ navigationController: UINavigationController, animationControllerFor operation: UINavigationController.Operation, from fromVC: UIViewController, to toVC: UIViewController) -> UIViewControllerAnimatedTransitioning? {
         if [fromVC, toVC].filter({$0 is MLCardFormViewController || $0 is PXSecurityCodeViewController}).count > 0 {
-        //if fromVC is MLCardFormViewController || toVC is MLCardFormViewController || toVC is PXSecurityCodeViewController {
             return PXOneTapViewControllerTransition()
         }
         return nil
